Minor refactoring to D2D compiler files
