+++ src/ComputeSharp.D2D1/Shaders/Translation/Headers/D2D1EffectHelpers.cs
